feat: create cardData module for English and Japanese sets; refactor imports in home and sets pages
