+++ src/pages/home/card-data.ts
@@ -1,0 +1,30 @@
+export const cardData = [
+	{
+		title: "English",
+		sets: [
+			"SV08: Surging Sparks",
+			"SV07: Stellar Crown",
+			"SV06: Twilight Masquerade",
+			"SV05: Temporal Forces",
+			"SV04: Paradox Rift",
+			"SV03: Obsidian Flames",
+			"SV: Shrouded Fable",
+			"SV: Scarlet & Violet 151",
+			"SV: Paldean Fates",
+		],
+	},
+	{
+		title: "Japanese",
+		sets: [
+			"SV7A: Paradise Dragona",
+			"SV7: Stellar Miracle",
+			"SV6A: Night Wanderer",
+			"SV6: Transformation Mask",
+			"SV5M: Cyber Judge",
+			"SV5K: Wild Force",
+			"SV5A: Crimson Haze",
+			"SV-P Promotional Cards",
+			"SV: Ancient Koraidon ex Starter Deck & Build Set",
+		],
+	},
+]
